+++ src/kotlin-intellij/main/uk/co/reecedunn/intellij/plugin/core/navigation/ItemPresentationEx.kt
@@ -18,5 +18,12 @@ package uk.co.reecedunn.intellij.plugin.core.navigation
 import com.intellij.navigation.ItemPresentation
 
 interface ItemPresentationEx : ItemPresentation {
-    val structurePresentableText: String?
+    enum class Type {
+        Default,
+        StructureView,
+        NavBar,
+        NavBarPopup
+    }
+
+    fun getPresentableText(type: Type): String?
 }